# swagger.yml
openapi: 3.0.0
info:
  title: "RP Flask REST API"
  description: "An API about todo Application"
  version: "1.0.0"
servers:
  - url: "/api"
components:
  schemas:
...
    EditStatus:
      type: object
      required:
        - username
        - task_name
        - subtask_title
        - _status
      properties:
        username:
          type: string
        task_name:
          type: string
        subtask_title:
          type: string
        _status:
          type: string
          enum:
            - unassigned
            - assigned
            - started
            - in-progress
            - done
            - null
...
    SubtaskData:
      type: object
      properties:
        username:
          type: string
        subtask_title:
          type: string
        task_name:
          type: string
        due_date:
          type: string
        content:
          type: string
      required:
        - username
        - task_name
        - subtask_title
        - due_date
    EditSubtaskData:
      type: object
      properties:
        username:
          type: string
        old_subtask_title:
          type: string
        new_subtask_title:
          type: string
        new_due_date:
          type: string
        task_name:
          type: string
        new_content:
          type: string
      required:
        - username
        - old_subtask_title
        - task_name
...
  securitySchemes:
    jwt_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: jwt_manager.security_handler
paths:
...
  /user/task/subtask/create:
    post:
      operationId: subtasks.create_subtask
      tags:
        - Subtask
      summary: "creates a subtask"
      requestBody:
        description: "subtask Data to create a task subtask"
        required: True
        content:
          application/x-www-form-urlencoded:
            schema:
              x-body-name: "subtask"
              $ref: "#/components/schemas/SubtaskData"
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: User SubTask has been successfully created
  /user/tasks/subtask/update:
    put:
      operationId: subtasks.update_subtask
      tags:
        - Subtask
      summary: "updates a subtask"
      requestBody:
        description: "subtask Data to create a task subtask"
        required: True
        content:
          application/x-www-form-urlencoded:
            schema:
              x-body-name: "subtask"
              $ref: "#/components/schemas/EditSubtaskData"
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: User SubTask has been successfully updated
  /user/{username}/task/{task_name}/subtask/{subtask_title}/get:
    get:
      operationId: subtasks.get_subtask
      tags:
        - Subtask
      summary: "Retrieves a user subtask"
      parameters:
        - name: username
          in: path
          description: The username of the user that created the subtask
          required: true
          schema:
            type: string
        - name: task_name
          in: path
          description: The name of the task
          required: true
          schema:
            type: string
        - name: subtask_title
          in: path
          description: The name of the subtask
          required: true
          schema:
            type: string
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: subtask Retrieved successfully
  /user/{username}/tasks/{task_name}/subtask/all:
    get:
      operationId: subtasks.get_all_subtasks
      tags:
        - Subtask
      summary: "Retrieves all user subtask"
      parameters:
        - name: username
          in: path
          description: The username of the user that created the subtasks
          required: true
          schema:
            type: string
        - name: task_name
          in: path
          description: The name of the task
          required: true
          schema:
            type: string
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: all subtasks Retrieved successfully
  /user/{username}/task/{task_name}/subtask/{subtask_title}/delete:
    delete:
      operationId: subtasks.delete_subtask
      tags:
        - Subtask
      summary: "Deletes a user subtask"
      parameters:
        - name: username
          in: path
          description: The username of the user that created the subtask
          required: true
          schema:
            type: string
        - name: task_name
          in: path
          description: The name of the task
          required: true
          schema:
            type: string
        - name: subtask_title
          in: path
          description: The title of the subtask
          required: true
          schema:
            type: string
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: subtask Delete successfully
  /user/tasks/{task_name}/subtask/{subtask_title}/assign/{assign_to}:
    put:
      operationId: subtasks.assign_subtask
      tags:
        - Subtask Assign
      summary: "Assigns a user subtask"
      parameters:
        - name: assign_to
          in: path
          description: The name of the user to be assigned subtask
          required: true
          schema:
            type: string
        - name: task_name
          in: path
          description: The name of the task
          required: true
          schema:
            type: string
        - name: subtask_title
          in: path
          description: The title of the subtask
          required: true
          schema:
            type: string
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: subtask assigned successfully
  /user/tasks/{task_name}/subtask/{subtask_title}/unassign/{unassign_from}:
    put:
      operationId: subtasks.unassign_subtask
      tags:
        - Subtask Assign
      summary: "Unassign a user subtask"
      parameters:
        - name: unassign_from
          in: path
          description: The name of the user  to be unassigned subtask
          required: true
          schema:
            type: string
        - name: task_name
          in: path
          description: The name of the task
          required: true
          schema:
            type: string
        - name: subtask_title
          in: path
          description: The title of the subtask
          required: true
          schema:
            type: string
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: subtask assigned successfully
  /user/task/subtask/status:
    put:
      operationId: subtasks.update_subtask_status
      tags:
        - Update Status
      summary: "Updates a user subtask status"
      requestBody:
        description: "subtask Data to update a task subtask status"
        required: True
        content:
          application/x-www-form-urlencoded:
            schema:
              x-body-name: "update status"
              $ref: "#/components/schemas/EditStatus"
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: subtask status updated successfully
...
  /user/tasks/{task_name}/subtask/{subtask_title}/assigned/all:
    get:
      operationId: subtasks.show_assigned_users_for_subtask
      tags:
        - Subtask Assign
      summary: "Retrieves a list of assigned users for a subtask"
      parameters:
        - name: task_name
          in: path
          description: The name of the task
          required: true
          schema:
            type: string
        - name: subtask_title
          in: path
          description: The title of the subtask
          required: true
          schema:
            type: string
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: list of assigned users retrieved successfully
...